added exception handler , switched from class to record in some of the dtos, moved jwt service
