+++ src/main/java/pl/kj/m33t/m33tbackend/model/entity/modules/Post.java
@@ -1,6 +1,7 @@
 package pl.kj.m33t.m33tbackend.model.entity.modules;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
