+++ src/main/java/pl/kj/m33t/m33tbackend/rest/EventRestControllerApi.java
@@ -2,9 +2,9 @@ package pl.kj.m33t.m33tbackend.rest;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import pl.kj.m33t.m33tbackend.model.entity.Event;
 import pl.kj.m33t.m33tbackend.service.dto.request.EventRequest;
 import pl.kj.m33t.m33tbackend.service.dto.response.EventResponse;
-import pl.kj.m33t.m33tbackend.model.entity.Event;
 
 import java.util.List;
