+++ src/main/java/pl/kj/m33t/m33tbackend/rest/impl/modules/DebtCalulatorController.java
@@ -3,10 +3,10 @@ package pl.kj.m33t.m33tbackend.rest.impl.modules;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import pl.kj.m33t.m33tbackend.model.entity.modules.DebtCalculator;
 import pl.kj.m33t.m33tbackend.service.dto.request.modules.DebtCalculatorRequest;
 import pl.kj.m33t.m33tbackend.service.dto.response.modules.DebtCalculatorResponse;
 import pl.kj.m33t.m33tbackend.service.mapping.modules.DebtCalculatorMapper;
-import pl.kj.m33t.m33tbackend.model.entity.modules.DebtCalculator;
 import pl.kj.m33t.m33tbackend.service.modules.DebtCalculatorService;
 
 @RestController
