+++ src/main/java/pl/kj/m33t/m33tbackend/rest/impl/modules/PostRestController.java
@@ -3,9 +3,9 @@ package pl.kj.m33t.m33tbackend.rest.impl.modules;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import pl.kj.m33t.m33tbackend.service.dto.response.modules.PostResponse;
-import pl.kj.m33t.m33tbackend.service.dto.request.modules.PostRequest;
 import pl.kj.m33t.m33tbackend.model.entity.modules.Post;
+import pl.kj.m33t.m33tbackend.service.dto.request.modules.PostRequest;
+import pl.kj.m33t.m33tbackend.service.dto.response.modules.PostResponse;
 import pl.kj.m33t.m33tbackend.service.mapping.modules.PostMapper;
 import pl.kj.m33t.m33tbackend.service.modules.PostService;
