+++ src/main/java/pl/kj/m33t/m33tbackend/rest/impl/modules/SurveyRestController.java
@@ -4,11 +4,11 @@ package pl.kj.m33t.m33tbackend.rest.impl.modules;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import pl.kj.m33t.m33tbackend.model.entity.modules.Survey;
 import pl.kj.m33t.m33tbackend.service.dto.request.modules.SurveyRequest;
 import pl.kj.m33t.m33tbackend.service.dto.response.modules.SurveyResponse;
-import pl.kj.m33t.m33tbackend.model.entity.modules.Survey;
-import pl.kj.m33t.m33tbackend.service.modules.SurveyService;
 import pl.kj.m33t.m33tbackend.service.mapping.modules.SurveyMapper;
+import pl.kj.m33t.m33tbackend.service.modules.SurveyService;
 
 @RestController
 @RequiredArgsConstructor
