+++ src/main/java/pl/kj/m33t/m33tbackend/service/AuthenticationService.java
@@ -5,13 +5,12 @@ import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import pl.kj.m33t.m33tbackend.config.JwtService;
-import pl.kj.m33t.m33tbackend.service.dto.request.AuthenticationRequest;
-import pl.kj.m33t.m33tbackend.service.dto.response.AuthenticationResponse;
-import pl.kj.m33t.m33tbackend.model.repository.UserRepository;
 import pl.kj.m33t.m33tbackend.model.entity.Role;
 import pl.kj.m33t.m33tbackend.model.entity.User;
+import pl.kj.m33t.m33tbackend.model.repository.UserRepository;
+import pl.kj.m33t.m33tbackend.service.dto.request.AuthenticationRequest;
 import pl.kj.m33t.m33tbackend.service.dto.request.RegisterRequest;
+import pl.kj.m33t.m33tbackend.service.dto.response.AuthenticationResponse;
 
 @Service
 @RequiredArgsConstructor
@@ -22,15 +21,13 @@ public class AuthenticationService {
     private final AuthenticationManager authenticationManager;
 
     public AuthenticationResponse register(RegisterRequest registerRequest){
-        User user = new User(1L,registerRequest.getNickname(),
-                registerRequest.getEmail(),
-                passwordEncoder.encode(registerRequest.getPassword()),
+        User user = new User(1L,registerRequest.nickname(),
+                registerRequest.email(),
+                passwordEncoder.encode(registerRequest.password()),
                 Role.USER);
         userRepository.save(user);
         String jwtToken = jwtService.generateToken(user);
-        return AuthenticationResponse.builder()
-                .token(jwtToken)
-                .build();
+        return new AuthenticationResponse(jwtToken);
     }
 
     public AuthenticationResponse authenticate(AuthenticationRequest authenticationRequest){
@@ -42,8 +39,6 @@ public class AuthenticationService {
         );
         User user = userRepository.findByEmail(authenticationRequest.getEmail()).orElseThrow();
         String jwtToken = jwtService.generateToken(user);
-        return AuthenticationResponse.builder()
-                .token(jwtToken)
-                .build();
+        return new AuthenticationResponse(jwtToken);
     }
 }
