+++ src/main/java/pl/kj/m33t/m33tbackend/service/dto/request/modules/SurveyRequest.java
@@ -1,6 +1,5 @@
 package pl.kj.m33t.m33tbackend.service.dto.request.modules;
 
-import java.util.HashMap;
 import java.util.List;
 
 public record SurveyRequest(String sentence, List<String> options) {}
