+++ src/main/java/pl/kj/m33t/m33tbackend/service/dto/response/AuthenticationResponse.java
@@ -1,11 +1,3 @@
 package pl.kj.m33t.m33tbackend.service.dto.response;
 
-import lombok.*;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class AuthenticationResponse {
-    private String token;
-}
+public record AuthenticationResponse(String token) {}
