+++ src/main/java/pl/kj/m33t/m33tbackend/service/dto/response/modules/SurveyResponse.java
@@ -1,7 +1,6 @@
 package pl.kj.m33t.m33tbackend.service.dto.response.modules;
 
 import java.util.Date;
-import java.util.HashMap;
 import java.util.Map;
 
 public record SurveyResponse(Long postId, Long eventId, String sentence, Map<String, Integer> optionsAndVotes, Date date) {}
