+++ src/main/java/pl/kj/m33t/m33tbackend/service/mapping/EventMapper.java
@@ -3,9 +3,9 @@ package pl.kj.m33t.m33tbackend.service.mapping;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.factory.Mappers;
+import pl.kj.m33t.m33tbackend.model.entity.Event;
 import pl.kj.m33t.m33tbackend.service.dto.request.EventRequest;
 import pl.kj.m33t.m33tbackend.service.dto.response.EventResponse;
-import pl.kj.m33t.m33tbackend.model.entity.Event;
 
 @Mapper
 public interface EventMapper {
