+++ src/main/java/pl/kj/m33t/m33tbackend/service/mapping/modules/PostMapper.java
@@ -3,9 +3,9 @@ package pl.kj.m33t.m33tbackend.service.mapping.modules;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.factory.Mappers;
+import pl.kj.m33t.m33tbackend.model.entity.modules.Post;
 import pl.kj.m33t.m33tbackend.service.dto.request.modules.PostRequest;
 import pl.kj.m33t.m33tbackend.service.dto.response.modules.PostResponse;
-import pl.kj.m33t.m33tbackend.model.entity.modules.Post;
 
 @Mapper
 public interface PostMapper {
