+++ src/main/java/pl/kj/m33t/m33tbackend/service/modules/DebtCalculatorService.java
@@ -2,14 +2,14 @@ package pl.kj.m33t.m33tbackend.service.modules;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import pl.kj.m33t.m33tbackend.model.repository.modules.DebtCalculatorRepository;
-import pl.kj.m33t.m33tbackend.model.repository.modules.PurchaseRepository;
-import pl.kj.m33t.m33tbackend.service.dto.request.modules.DebtCalculatorRequest;
 import pl.kj.m33t.m33tbackend.model.entity.Event;
 import pl.kj.m33t.m33tbackend.model.entity.modules.DebtCalculator;
-import pl.kj.m33t.m33tbackend.service.mapping.modules.DebtCalculatorMapper;
+import pl.kj.m33t.m33tbackend.model.repository.modules.DebtCalculatorRepository;
+import pl.kj.m33t.m33tbackend.model.repository.modules.PurchaseRepository;
 import pl.kj.m33t.m33tbackend.service.EventService;
 import pl.kj.m33t.m33tbackend.service.UserService;
+import pl.kj.m33t.m33tbackend.service.dto.request.modules.DebtCalculatorRequest;
+import pl.kj.m33t.m33tbackend.service.mapping.modules.DebtCalculatorMapper;
 
 @Service
 @RequiredArgsConstructor
