+++ src/main/java/pl/kj/m33t/m33tbackend/service/modules/SurveyService.java
@@ -4,14 +4,14 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import pl.kj.m33t.m33tbackend.service.dto.request.modules.SurveyRequest;
 import pl.kj.m33t.m33tbackend.model.entity.Event;
 import pl.kj.m33t.m33tbackend.model.entity.User;
 import pl.kj.m33t.m33tbackend.model.entity.modules.Survey;
-import pl.kj.m33t.m33tbackend.service.EventService;
 import pl.kj.m33t.m33tbackend.model.repository.modules.SurveyRepository;
-import pl.kj.m33t.m33tbackend.service.mapping.modules.SurveyMapper;
+import pl.kj.m33t.m33tbackend.service.EventService;
 import pl.kj.m33t.m33tbackend.service.UserService;
+import pl.kj.m33t.m33tbackend.service.dto.request.modules.SurveyRequest;
+import pl.kj.m33t.m33tbackend.service.mapping.modules.SurveyMapper;
 
 @Service
 @RequiredArgsConstructor
