+++ src/test/java/pl/kj/m33t/m33tbackend/services/EventServiceImplTest.java
@@ -8,19 +8,18 @@ import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 import org.mockito.junit.MockitoJUnitRunner;
 import org.springframework.security.core.context.SecurityContextHolder;
-import pl.kj.m33t.m33tbackend.model.repository.EventRepository;
 import pl.kj.m33t.m33tbackend.model.entity.Event;
 import pl.kj.m33t.m33tbackend.model.entity.EventType;
 import pl.kj.m33t.m33tbackend.model.entity.Role;
 import pl.kj.m33t.m33tbackend.model.entity.User;
 import pl.kj.m33t.m33tbackend.model.entity.modules.EventModule;
+import pl.kj.m33t.m33tbackend.model.repository.EventRepository;
 import pl.kj.m33t.m33tbackend.service.EventServiceImpl;
 import pl.kj.m33t.m33tbackend.service.UserService;
 
 import java.util.*;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.Mockito.*;
 
 @RunWith(MockitoJUnitRunner.class)
