+++ src/test/java/pl/kj/m33t/m33tbackend/services/modules/PostServiceTest.java
@@ -6,20 +6,17 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 import pl.kj.m33t.m33tbackend.model.entity.Event;
+import pl.kj.m33t.m33tbackend.model.entity.modules.Post;
+import pl.kj.m33t.m33tbackend.model.repository.modules.PostRepository;
+import pl.kj.m33t.m33tbackend.service.EventService;
+import pl.kj.m33t.m33tbackend.service.dto.request.modules.PostRequest;
+import pl.kj.m33t.m33tbackend.service.modules.PostService;
 
 import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.Mockito.*;
-
-import pl.kj.m33t.m33tbackend.model.repository.modules.PostRepository;
-import pl.kj.m33t.m33tbackend.service.dto.request.modules.PostRequest;
-import pl.kj.m33t.m33tbackend.model.entity.modules.Post;
-import pl.kj.m33t.m33tbackend.service.EventService;
-import pl.kj.m33t.m33tbackend.service.modules.PostService;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.mockito.Mockito.*;
 
 public class PostServiceTest {
